+++ test/test_helper.rb
@@ -9,36 +9,23 @@ bundle exec rake db:migrate RAILS_ENV=test
 =end
 
 require 'simplecov'
+require 'minitest/pride'
 
-
-ENV["RAILS_ENV"] ||= "test"
-require File.expand_path("../../config/environment", __FILE__)
-require "rails/test_help"
-require "minitest/rails"
-
-# To add Capybara feature tests add `gem "minitest-rails-capybara"`
-# to the test group in the Gemfile and uncomment the following:
-require "minitest/rails/capybara"
-
-# Uncomment for awesome colorful output
-require "minitest/pride"
+ENV['RAILS_ENV'] ||= 'test'
+require File.expand_path('../../config/environment', __FILE__)
+require 'rails/test_help'
+require 'minitest/rails'
 
 class ActiveSupport::TestCase
-  setup :global_setup
-
-  ActiveRecord::Migration.check_pending!
-
-  # Setup all fixtures in test/fixtures/*.(yml|csv) for all tests in alphabetical order.
-  #
-  # Note: You'll currently still have to declare fixtures explicitly in integration tests
-  # -- they do not yet inherit this setting
+  # Setup all fixtures in test/fixtures/*.yml for all tests in alphabetical order.
   fixtures :all
 
-  def global_setup
+  # Add more helper methods to be used by all tests here...
+  setup :global_setup
 
+  def global_setup
     Survey.refresh_all_surveys
   end
-  # Add more helper methods to be used by all tests here...
 end
 
 class ActionController::TestCase
@@ -46,14 +33,13 @@ class ActionController::TestCase
 
   def login(resource)
     @request.env["devise.mapping"] = Devise.mappings[resource]
-    sign_in(:user, resource)
+    sign_in(resource.class.name.downcase.to_sym, resource)
   end
 
   def assert_authorization_exception
     assert_response 302
     assert flash[:alert]
   end
-
 end
 
 class ActionDispatch::IntegrationTest
@@ -61,7 +47,18 @@ class ActionDispatch::IntegrationTest
     user = User.create(password: password, password_confirmation: password, email: email,
                        first_name: user_template.first_name, last_name: user_template.last_name)
     user.save!
-    post_via_redirect '/users', user: { email: email, password: password }
+    user.update_column :deleted, user_template.deleted?
+    post_via_redirect '/users/sign_in', user: { email: email, password: password }
     user
+  end
+end
+
+module Rack
+  module Test
+    class UploadedFile
+      def tempfile
+        @tempfile
+      end
+    end
   end
 end